00029 - Refactor AutocompleteEditableStringField to use EditableFieldFrame for improved editing experience; enhance InputLabel with size support; update AutocompleteInput to autoFocus; adjust DealDetails to use larger size for better UI consistency.
