+++ src/components/common/inputs/EditableStringField.tsx
@@ -71,7 +71,7 @@ export default function EditableStringField({
       }: EditableFieldFrameChildProps): React.ReactNode {
         return (
           <input
-            className={`input border rounded px-2 py-1 w-full block truncate`}
+            className={"input border rounded px-2 py-1 w-full block truncate"}
             type="text"
             value={inputValue as string}
             onChange={(e) => setInputValue(e.target.value)}
