+++ src/components/common/inputs/InputLabel.tsx
@@ -1,7 +1,17 @@
-export default function InputLabel({ label }: { label?: string }) {
+import { SizeType } from "../sizeUtils";
+
+export default function InputLabel({
+  label,
+  size,
+}: {
+  label?: string;
+  size?: SizeType;
+}) {
   if (!label) return null;
   return (
-    <label className="mt- mr-2 text-sm font-semibold text-gray-600 dark:text-gray-400 whitespace-nowrap">
+    <label
+      className={`mr-2 text-${size || "sm"} font-semibold text-gray-600 dark:text-gray-400 whitespace-nowrap`}
+    >
       {label}:
     </label>
   );
